Implemented the functionality to allow users to create new accounts. I added a bunch of code in there for how to properly use session variables, hash passwords, send commands to SQLite, and so on.
I additionally added a generalFunctions.php file, which I'm intending to be a place where we can store functions which are used over and over throughout different files, to reduce code repetition.

I added a <noscript> tag to the register.php page. We need to make sure the same tag gets placed on each page.

I added a /js/ directory in which javascript files should go. Feel free to make subdirectories if we need more organization.