Implemented the ability to log in to an existing account through the /login.php page.
Fixed the password hashing; I was doing it completely wrong.
The PHPass implementation we're using automatically takes care of password salts, so we don't need to store them ourselves.
I've updated the database to reflect this change. If you're already built yours, you'll need to build it again using buildDatabase.sql

Made the requirements for one's username more strict---can only contain letters, #s, and underscores.
This is partly because we'll be using the username as a GET variable later.